GC_HACKATHON/cricket_pipeline/requirements.txt
--prefer-binary
pathway
litellm==1.40.0
openai
google-generativeai
numpy
fastapi
uvicorn
werkzeug
serpapi
beautifulsoup4
requests
aiohttp
python-dotenv
pytesseract
python-multipart
langchain
langchain_openai
openai
flask
phidata
google-generativeai
duckduckgo_search
groq